API/requirements.txt
fastapi~=0.110.2
pydantic~=2.7.0
python-dotenv~=1.0.1
g4f~=0.3.0.8
uvicorn~=0.29.0